# docs/14_OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: Enterprise AI Operations Center API
  description: |
    Unified API for the Enterprise AI Operations Center — Multi-Agent Orchestration,
    Secure RAG, Multimodal Understanding, Voice Interface, Edge Deployment,
    Enterprise RBAC Security, and MLOps Observability.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@eaioc.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.eaioc.example.com/api/v1
    description: Production
  - url: https://api-staging.eaioc.example.com/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local Development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      operationId: registerUser
      summary: Register new user and organization
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      tags: [Authentication]
      operationId: loginUser
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LoginResponse'
                  - $ref: '#/components/schemas/MFARequiredResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: List agents
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
    post:
      tags: [Agents]
      operationId: createAgent
      summary: Create new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
...
  /workflows/{workflowId}/execute:
    post:
      tags: [Agents]
      operationId: executeWorkflow
      summary: Execute workflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteWorkflowRequest'
      responses:
        '202':
          description: Execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionCreatedResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/sso/callback
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from /users/me/api-keys
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Pagination cursor from previous response
      schema:
        type: string
    PageSize:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, full_name, organization_name]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 12
          maxLength: 128
        full_name:
          type: string
          maxLength: 255
        organization_name:
          type: string
          maxLength: 255
    RegisterResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user_id:
              type: string
              format: uuid
            email:
              type: string
            tenant_id:
              type: string
              format: uuid
            organization_name:
              type: string
            is_verified:
              type: boolean
            message:
              type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            token_type:
              type: string
              enum: [Bearer]
            expires_in:
              type: integer
              description: Access token TTL in seconds
            user:
              $ref: '#/components/schemas/UserSummary'
    MFARequiredResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            mfa_required:
              type: boolean
              enum: [true]
            mfa_token:
              type: string
            message:
              type: string
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        tenant_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
...
    CreateAgentRequest:
      type: object
      required: [name, type, model_config]
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        type:
          type: string
          enum: [conversational, research, code_review, data_analysis, custom]
        model_config:
          $ref: '#/components/schemas/ModelConfig'
        system_prompt:
          type: string
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolConfig'
        guardrails:
          $ref: '#/components/schemas/Guardrails'
        cost_budget:
          $ref: '#/components/schemas/CostBudget'
...
    AgentResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            description:
              type: string
            version:
              type: string
            type:
              type: string
            status:
              type: string
            model_config:
              $ref: '#/components/schemas/ModelConfig'
            tools:
              type: array
              items:
                $ref: '#/components/schemas/ToolConfig'
            guardrails:
              $ref: '#/components/schemas/Guardrails'
            cost_budget:
              $ref: '#/components/schemas/CostBudget'
            created_by:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    AgentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AgentResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ModelConfig:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
          enum: [openai, anthropic, google, mistral, ollama, custom]
        model:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        max_tokens:
          type: integer
          minimum: 1
          maximum: 128000
          default: 4096
        top_p:
          type: number
          default: 1.0
        fallback_chain:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              model:
                type: string
    ToolConfig:
      type: object
      required: [name]
      properties:
        name:
          type: string
        config:
          type: object
    Guardrails:
      type: object
      properties:
        block_pii:
          type: boolean
          default: false
        pii_types:
          type: array
          items:
            type: string
        content_filter:
          type: string
          enum: [none, moderate, strict]
          default: moderate
        require_citations:
          type: boolean
          default: false
        max_output_tokens:
          type: integer
    CostBudget:
      type: object
      properties:
        max_cost_per_execution:
          type: number
          default: 1.0
        max_tokens_per_execution:
          type: integer
          default: 100000
...
    ExecuteWorkflowRequest:
      type: object
      properties:
        input_data:
          type: object
        options:
          type: object
          properties:
            async:
              type: boolean
              default: true
            webhook_url:
              type: string
              format: uri
    ExecutionCreatedResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            execution_id:
              type: string
              format: uuid
            workflow_id:
              type: string
              format: uuid
            status:
              type: string
            websocket_url:
              type: string
            created_at:
              type: string
              format: date-time
...
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        total_count:
          type: integer
        page_size:
          type: integer
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    Conflict:
      description: Resource conflict (e.g., duplicate email)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'